+++ src/services/app-events.service.ts
@@ -1,0 +1,5 @@
+import { Subject } from 'rxjs';
+
+export class AppEventsService {
+  public static readonly mouseOverSvgEditPoint$: Subject<number> = new Subject<number>();
+}