+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## unreleased
 
+Features:
+
+- Highlight svg edit point control and form point when mouse over them for path shapes.
+
+Fixes:
+
 - Avoid adding path commands (mainly Z) when typing in other inputs of the app and the shape is already finished.
 - Avoid navigating back and forward when moving a shape left or right using alt key.
 - Add scrolling to attributes and shape list components. Limit the max height of the attributes component when path contains a lot of commands.